+ Restructure code, implement interfaces, and refactor test classes
