+++ illyriacore/src/main/java/eu/illyrion/illyriacore/interfaces/PermissionsInterface.java
@@ -1,8 +1,8 @@
-package eu.illyrion.illyriacore.configs;
+package eu.illyrion.illyriacore.interfaces;
 
-public class Permissions {
+public interface PermissionsInterface {
 
+    // General
     public static final String PERM_PREFIX = "illyriautils.";
     public static final String RELOAD = PERM_PREFIX + "reload";
-
 }
